:ambulance: :: interface로 Service로직 구현 빼기
