+++ src/main/kotlin/com/msg/gcms/domain/clubMember/presentation/ClubMemberController.kt
@@ -2,6 +2,7 @@ package com.msg.gcms.domain.clubMember.presentation
 
 import com.msg.gcms.domain.clubMember.presentation.data.response.ClubMemberListResponse
 import com.msg.gcms.domain.clubMember.service.FindClubMemberListService
+import com.msg.gcms.domain.clubMember.service.impl.FindClubMemberListServiceImpl
 import com.msg.gcms.domain.clubMember.util.ClubMemberConverter
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.*
